stok işlemleri ile iligli kısım tamamlandı.

+++ ProductionApp.DTOs/StockDtos/StockUpdateDto.cs
@@ -9,7 +9,7 @@ public class StockUpdateDto : IBaseDto
     public string? Name { get; set; }
     public string? Description { get; set; }
     public decimal Price { get; set; }
-    public decimal VAT { get; set; }
+    public int VAT { get; set; }
     public DateTime? ModifiedDate { get; set; }
     public int GroupId { get; set; }
 }
